cloud/api/requirements.txt
# CelesteOS Cloud API Requirements
# Python 3.11+

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Supabase client
supabase==2.3.0

# JWT authentication
PyJWT==2.8.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0

# HTTP and async
httpx==0.25.2
aiofiles==23.2.1

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring (optional)
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1
